requirements.txt
temporalio>=1.8.0
fastmcp>=0.4.0
httpx>=0.27.0
openai>=1.35.0
python-dotenv>=1.0.1
numpy>=1.26.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
streamlit>=1.35.0
openai-agents
crewai